requirements.txt
# ── Web Framework ──────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── LangChain Ecosystem (FREE) ─────────────────
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-openai
langgraph
langsmith

# ── Embeddings (FREE — runs locally on CPU) ─────
sentence-transformers==3.0.1

# ── Vector Store (FREE — local persistent) ──────
faiss-cpu==1.8.0.post1

# ── Validation ─────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Environment ────────────────────────────────
python-dotenv==1.0.1

# ── HTTP client ────────────────────────────────
httpx==0.27.0

# ── Token counting (RAG chunker) ───────────────
tiktoken==0.7.0
